Refactoring localization
We now detect the language using the `window.navigator` object, and webpack now splits the translations into their own resources that are loaded as needed per detected language. Language direction is now detected, too.

Note: this uses `React.Suspend`, which is an experimental feature, but in my opinion can be used in the exact case here, where we are simply delaying rendering until a Promise gets resolved (hence the return of a `throw`, which interacts with the `ErrorBoundary` of `React.Suspend`).