Integrating react-intl and Transifex (#1)
The main source file for our langauge strings is now generated by using the formatjs cli tool in `npm run transifex:extract`. This creates a `localization/transifex/source.json` file that can be uploaded to transifex using `npm run transifex:upload`. Downloading existing translations is supported with `npm run:transifex:download`.

On a technical note, refactored internationalization to use `react-intl`'s new `useIntl()` hook for injection into low-level properties such as input labels. This has a good side effect of not needing to pass `locale` around anymore. `messages.js` (always meant to be temporary) is now gone in favor of extracting source strings directly via static analysis of components in code.

The <Filter> component accepts localized strings, so it's been added as a special extra component for formatjs-cli to look for during the `npm run transifex:extract` step.

We switch to the new babel monorepo eslint parser, the old `babel-eslint` is deprecated and was throwing errors on valid code.

We now detect the language using the `window.navigator` object, and webpack now splits the translations into their own resources that are loaded as needed per detected language. Language direction is now detected, too.

Note: this uses `React.Suspend`, which is an experimental feature, but in my opinion can be used in the exact case here, where we are simply delaying rendering until a Promise gets resolved (hence the return of a `throw`, which interacts with the `ErrorBoundary` of `React.Suspend`).

The Filter component now takes a `header` prop of `FormattedMessage`. This gives us the flexibility to pass multiple internationalized strings to the Filter in the future as needed, and removes the need to tag Filter as a special component for the `formatjs` parser to scan for.