Externalize config.js
The `config.js` is now copied into `dist/` on build and included as an external `<script>` resource in `index.html`. The `externals` directive in the webpack config tells webpack that the `import` of `'config'` in `src/index.js` should simply expect the `config` variable to already be loaded in the application through the `<script>` tag.