Webpack perf and CI integration (#2)
* Split out a vendor bundle for Webpack

This means that as long as vendor libraries don't change, the client will cache the libraries, only needing to load new application code when it changes.

* Externalize config.js

The `config.js` is now copied into `dist/` on build and included as an external `<script>` resource in `index.html`. The `externals` directive in the webpack config tells webpack that the `import` of `'config'` in `src/index.js` should simply expect the `config` variable to already be loaded in the application through the `<script>` tag.

* Speed up webpack build

Previously `babel-loader` was accidentally parsing in `node_modules` on build. This is no longer the case, and build times have improved by 50%.

* Slack integration in CI