Adding content type filtering
Refactoring the `Filter` component to accept arrays of object values in addition to a JSONAPI query. If a query is not specified as a `query` prop and an array is present in the `value` prop, a list of checkboxes is rendered of the values passed in. The values are passed up to `App` and back down into `Search`.